+++ src/main/java/fr/erickfranco/cv_api/repositories/ExperienceRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ExperienceRepository extends JpaRepository<Experience, Integer> {
+public interface ExperienceRepository extends JpaRepository<Experience, Long> {
 }
