+++ src/main/java/fr/erickfranco/cv_api/repositories/ProjetRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProjetRepository extends JpaRepository<Projet, Integer> {
+public interface ProjetRepository extends JpaRepository<Projet, Long> {
 
 }
